updated distance and run scripts and timings

+++ scripts/roseobacter_runs.py
@@ -7,9 +7,8 @@ import glob
 #
 # run as
 # python roseobacter_runs.py
-#   ../../data/roseobacter/
-#   ../../runs/roseobacter
-#   ../../build/arakawa.x ../../software/kmacs/kmacs ../../software/spaced-prdna/spaced
+#  ../data/roseobacter/  ../runs/roseobacter
+#  ../build/arakawa.x ../software/kmacs/kmacs ../software/spaced-prdna/spaced
 
 in_dir = os.path.abspath(sys.argv[1])
 out_dir = os.path.abspath(sys.argv[2])
@@ -22,6 +21,8 @@ for k in [0, 1, 2, 3, 4, 5]:
     for i, fa_name in enumerate(fas_lst):
         out_file = os.path.join(out_dir, os.path.basename(fa_name))
         out_file = out_file.replace("fas", "acs.k" + str(k) + ".out")
+        if os.path.isfile(out_file):
+            continue
         print i, k, fa_name, out_file
         rc = 0
         rc = subprocess.call([kacs_exe, "-f", fa_name, "-k", str(k),
@@ -30,11 +31,13 @@ for k in [0, 1, 2, 3, 4, 5]:
             print "ERROR!"
             sys.exit(1)
 
-for k in [2, 4, 6]:
+for k in [0, 2, 3, 4, 5, 6]:
     for i, fa_name in enumerate(fas_lst):
         out_file = os.path.join(out_dir, os.path.basename(fa_name))
         out_file = out_file.replace("fas", "kmacs.k" + str(k) + ".out")
         log_file = out_file + ".log"
+        if os.path.isfile(out_file):
+            continue
         print i, k, fa_name, out_file
         rc = 0
         with open(log_file, "w") as tfo:
@@ -48,6 +51,8 @@ for i, fa_name in enumerate(fas_lst):
     out_file = os.path.join(out_dir, os.path.basename(fa_name))
     out_file = out_file.replace("fas", "spaced.out")
     log_file = out_file + ".log"
+    if os.path.isfile(out_file):
+        continue
     print i, 0, fa_name, out_file
     rc = 0
     with open(log_file, "w") as tfo:
